Refactor profile context: separate context definition and provider into distinct files
